31/7/21

+++ institute-management-spring/instititutemanagement/src/main/java/com/usecase/instititutemanagement/model/Student.java
@@ -1,9 +1,28 @@
 package com.usecase.instititutemanagement.model;
 import java.io.Serializable;
 
+import java.util.Date;
+import java.util.List;
+
+import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.ManyToMany;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
+import javax.persistence.Transient;
+import javax.persistence.UniqueConstraint;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
 
 @Entity
 @Table(name = "tbl_student")
@@ -124,11 +143,11 @@ public class Student implements Serializable {
 		this.surname = surname;
 	}
 
-	public int getMobilePhone() {
+	public long getMobilePhone() {
 		return mobilePhone;
 	}
 
-	public void setMobilePhone(int mobilePhone) {
+	public void setMobilePhone(Long mobilePhone) {
 		this.mobilePhone = mobilePhone;
 	}
 	public String getCampus() {
